+++ modules/adfBidAdapter.ts
@@ -7,6 +7,74 @@ import type { AdapterRequest, AdapterResponse, BidderSpec, ExtendedResponse, Ser
 import type { BidRequest, ClientBidderRequest } from '../src/adapterManager.js';
 import type { Bid } from '../src/bidfactory.js';
 
+/**
+ * Common optional parameters shared by all Adf bid request configurations.
+ */
+interface AdfCommonParams {
+  /**
+   * Ad exchange domain. Defaults to `'adx.adform.net'`.
+   */
+  adxDomain?: string;
+  /**
+   * Price type for bid responses: `'net'` or `'gross'`. Defaults to `'net'`.
+   */
+  pt?: 'net' | 'gross';
+  /**
+   * @deprecated Use `pt` instead.
+   */
+  priceType?: 'net' | 'gross';
+}
+
+/**
+ * Configuration using a master tag ID.
+ */
+interface AdfMidParams extends AdfCommonParams {
+  /**
+   * Master tag ID on the Adform platform.
+   */
+  mid: string | number;
+  inv?: never;
+  mname?: never;
+}
+
+/**
+ * Configuration using an inventory source and master tag name.
+ */
+interface AdfInvParams extends AdfCommonParams {
+  mid?: never;
+  /**
+   * Inventory source ID on the Adform platform.
+   */
+  inv: number;
+  /**
+   * Master tag name on the Adform platform.
+   */
+  mname: string;
+}
+
+/**
+ * Bidder parameters for the Adf (Adform) adapter.
+ *
+ * Either `mid` or both `inv` and `mname` must be provided.
+ */
+export type AdfBidderParams = AdfMidParams | AdfInvParams;
+
+declare module '../src/adUnits' {
+  interface BidderParams {
+    adf: AdfBidderParams;
+    adform: AdfBidderParams;
+    adformOpenRTB: AdfBidderParams;
+  }
+}
+
+declare global {
+  interface Window {
+    Adform: {
+      renderOutstream(bid: Bid): void;
+    };
+  }
+}
+
 const BIDDER_CODE = 'adf';
 const GVLID = 50;
 const BIDDER_ALIAS = [
